// src/app/models/sedetraslado.model.ts
export class SedeTraslado {
    id: string;
    nombre: string;
    ubigeo: string;
    cod_ubigeo: string;
    departamento: string;
    provincia: string;
    distrito: string;
    direccion: string;
    referencia: string;
    ruc: string;
    usuario_creacion: string;
    usuario_modificacion: string;
}

export class SedeTrasladoFiltro {
    id: string;
    ruc: string;
    ubigeo: string;
    departamento: string;
    provincia: string;
    distrito: string;
    nombre: string;
    direccion: string;
    referencia: string;
}